Fixed new design of Asset for Address.
