+++ opencti-platform/opencti-front/src/private/components/assets/devices/DeviceCreationDetails.js
@@ -30,6 +30,7 @@ import { adaptFieldValue } from '../../../../utils/String';
 import ItemIcon from '../../../../components/ItemIcon';
 import InstalledAsset from '../../common/form/InstalledAsset';
 import PortsField from '../../common/form/PortsField';
+import AddressField from '../../common/form/AddressField';
 
 const styles = (theme) => ({
   drawerPaper: {
@@ -381,52 +382,6 @@ class DeviceCreationDetailsComponent extends Component {
                   fullWidth={true}
                 />
               </div>
-              <div>
-                <Typography
-                  variant="h3"
-                  color="textSecondary"
-                  gutterBottom={true}
-                  style={{ float: 'left', marginTop: 20 }}
-                >
-                  {t('IPv4 Address')}
-                </Typography>
-                <div style={{ float: 'left', margin: '21px 0 0 5px' }}>
-                  <Tooltip title={t('ipv4_address')} >
-                    <Information fontSize="inherit" color="disabled" />
-                  </Tooltip>
-                </div>
-                <Field
-                  component={TextField}
-                  variant='outlined'
-                  name="ipv4_address"
-                  size='small'
-                  multiline={true}
-                  fullWidth={true}
-                />
-              </div>
-              <div>
-                <Typography
-                  variant="h3"
-                  color="textSecondary"
-                  gutterBottom={true}
-                  style={{ float: 'left', marginTop: 20 }}
-                >
-                  {t('IPv6 Address')}
-                </Typography>
-                <div style={{ float: 'left', margin: '21px 0 0 5px' }}>
-                  <Tooltip title={t('ipv6_address')} >
-                    <Information fontSize="inherit" color="disabled" />
-                  </Tooltip>
-                </div>
-                <Field
-                  component={TextField}
-                  variant='outlined'
-                  name="ipv6_address"
-                  size='small'
-                  multiline={true}
-                  fullWidth={true}
-                />
-              </div>
             </Grid>
             <Grid item={true} xs={6}>
               <div>
@@ -500,28 +455,6 @@ class DeviceCreationDetailsComponent extends Component {
                   component={TextField}
                   variant='outlined'
                   name="model"
-                  size='small'
-                  fullWidth={true}
-                />
-              </div>
-              <div>
-                <Typography
-                  variant="h3"
-                  color="textSecondary"
-                  gutterBottom={true}
-                  style={{ float: 'left', marginTop: 20 }}
-                >
-                  {t('MAC Address')}
-                </Typography>
-                <div style={{ float: 'left', margin: '21px 0 0 5px' }}>
-                  <Tooltip title={t('MAC Address')}>
-                    <Information fontSize="inherit" color="disabled" />
-                  </Tooltip>
-                </div>
-                <Field
-                  component={TextField}
-                  variant='outlined'
-                  name="mac_address"
                   size='small'
                   fullWidth={true}
                 />
@@ -663,6 +596,33 @@ class DeviceCreationDetailsComponent extends Component {
                   fullWidth={true}
                 />
               </div>
+            </Grid>
+            <Grid item={true} xs={12}>
+              <AddressField
+                setFieldValue={setFieldValue}
+                values={values}
+                addressValues={values.mac_address}
+                title='Mac Address'
+                name='mac_address'
+              />
+            </Grid>
+            <Grid item={true} xs={12}>
+              <AddressField
+                setFieldValue={setFieldValue}
+                values={values}
+                addressValues={values.ipv4_address}
+                title='IPv4 Address'
+                name='ipv4_address'
+              />
+            </Grid>
+            <Grid item={true} xs={12}>
+              <AddressField
+                setFieldValue={setFieldValue}
+                values={values}
+                addressValues={values.ipv6_address}
+                title='IPv6 Address'
+                name='ipv6_address'
+              />
             </Grid>
           </Grid>
         </Paper>
