+++ opencti-platform/opencti-front/src/private/components/assets/devices/DeviceEditionDetails.js
@@ -1,5 +1,3 @@
-/* eslint-disable */
-/* refactor */
 import React, { Component } from 'react';
 import * as PropTypes from 'prop-types';
 import graphql from 'babel-plugin-relay/macro';
@@ -30,6 +28,7 @@ import { adaptFieldValue } from '../../../../utils/String';
 import InstalledAsset from '../../common/form/InstalledAsset';
 import ItemIcon from '../../../../components/ItemIcon';
 import PortsField from '../../common/form/PortsField';
+import AddressField from '../../common/form/AddressField';
 
 const styles = (theme) => ({
   paper: {
@@ -334,52 +333,6 @@ class DeviceEditionDetailsComponent extends Component {
                     fullWidth={true}
                   />
                 </div>
-                <div>
-                  <Typography
-                    variant="h3"
-                    color="textSecondary"
-                    gutterBottom={true}
-                    style={{ float: 'left', marginTop: 20 }}
-                  >
-                    {t('IPv4 Address')}
-                  </Typography>
-                  <div style={{ float: 'left', margin: '21px 0 0 5px' }}>
-                    <Tooltip title={t('ipv4_address')} >
-                      <Information fontSize="inherit" color="disabled" />
-                    </Tooltip>
-                  </div>
-                  <Field
-                    component={TextField}
-                    variant='outlined'
-                    name="ipv4_address"
-                    size='small'
-                    multiline={true}
-                    fullWidth={true}
-                  />
-                </div>
-                <div>
-                  <Typography
-                    variant="h3"
-                    color="textSecondary"
-                    gutterBottom={true}
-                    style={{ float: 'left', marginTop: 20 }}
-                  >
-                    {t('IPv6 Address')}
-                  </Typography>
-                  <div style={{ float: 'left', margin: '21px 0 0 5px' }}>
-                    <Tooltip title={t('ipv6_address')} >
-                      <Information fontSize="inherit" color="disabled" />
-                    </Tooltip>
-                  </div>
-                  <Field
-                    component={TextField}
-                    variant='outlined'
-                    name="ipv6_address"
-                    size='small'
-                    multiline={true}
-                    fullWidth={true}
-                  />
-                </div>
               </Grid>
               <Grid item={true} xs={6}>
                 <div>
@@ -411,27 +364,27 @@ class DeviceEditionDetailsComponent extends Component {
                   />
                 </div>
                 <div>
-                <Typography
-                  variant="h3"
-                  color="textSecondary"
-                  gutterBottom={true}
-                  style={{ float: 'left', marginTop: 15 }}
-                >
-                  {t('Location')}
-                </Typography>
-                <div style={{ float: 'left', margin: '16px 0 0 5px' }}>
-                  <Tooltip title={t('Location')}>
-                    <Information fontSize="inherit" color="disabled" />
-                  </Tooltip>
-                </div>
-                <div className="clearfix" />
-                <Field
-                  component={TextField}
-                  name="locations"
-                  fullWidth={true}
-                  multiline={true}
-                  rows="3"
-                  variant='outlined'
+                  <Typography
+                    variant="h3"
+                    color="textSecondary"
+                    gutterBottom={true}
+                    style={{ float: 'left', marginTop: 15 }}
+                  >
+                    {t('Location')}
+                  </Typography>
+                  <div style={{ float: 'left', margin: '16px 0 0 5px' }}>
+                    <Tooltip title={t('Location')}>
+                      <Information fontSize="inherit" color="disabled" />
+                    </Tooltip>
+                  </div>
+                  <div className="clearfix" />
+                  <Field
+                    component={TextField}
+                    name="locations"
+                    fullWidth={true}
+                    multiline={true}
+                    rows="3"
+                    variant='outlined'
                   />
                 </div>
                 <div>
@@ -452,32 +405,6 @@ class DeviceEditionDetailsComponent extends Component {
                     component={TextField}
                     variant='outlined'
                     name="model"
-                    size='small'
-                    fullWidth={true}
-                  />
-                </div>
-                <div>
-                  <Typography
-                    variant="h3"
-                    color="textSecondary"
-                    gutterBottom={true}
-                    style={{ float: 'left', marginTop: 20 }}
-                  >
-                    {t('MAC Address')}
-                  </Typography>
-                  <div style={{ float: 'left', margin: '21px 0 0 5px' }}>
-                    <Tooltip
-                      title={t(
-                        'In OpenCTI, a predictable STIX ID is generated based on one or multiple attributes of the entity.',
-                      )}
-                    >
-                      <Information fontSize="inherit" color="disabled" />
-                    </Tooltip>
-                  </div>
-                  <Field
-                    component={TextField}
-                    variant='outlined'
-                    name="mac_address"
                     size='small'
                     fullWidth={true}
                   />
@@ -619,6 +546,33 @@ class DeviceEditionDetailsComponent extends Component {
                     fullWidth={true}
                   />
                 </div>
+              </Grid>
+              <Grid item={true} xs={12}>
+                <AddressField
+                  setFieldValue={setFieldValue}
+                  values={values}
+                  addressValues={values.mac_address}
+                  title='Mac Address'
+                  name='mac_address'
+                />
+              </Grid>
+              <Grid item={true} xs={12}>
+                <AddressField
+                  setFieldValue={setFieldValue}
+                  values={values}
+                  addressValues={values.ipv4_address}
+                  title='IPv4 Address'
+                  name='ipv4_address'
+                />
+              </Grid>
+              <Grid item={true} xs={12}>
+                <AddressField
+                  setFieldValue={setFieldValue}
+                  values={values}
+                  addressValues={values.ipv6_address}
+                  title='IPv6 Address'
+                  name='ipv6_address'
+                />
               </Grid>
             </Grid>
           </Paper>
